minor large-file fixes
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@15599 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/fitsioutils.h
@@ -194,7 +194,7 @@ int fits_offset_of_column(qfits_table* table, int colnum);
  */
 int fits_convert_data(void* dest, int deststride, tfits_type desttype,
                       const void* src, int srcstride, tfits_type srctype,
-                      int arraysize, int N);
+                      int arraysize, size_t N);
 
 // write single column fields:
 int fits_write_data_A(FILE* fid, char value);
